Moved app logic to PostsPage
